refactored app jsx and index js as requested by Thomas, new file created from refactoring called Display
