+++ src/components/DisplayHeader.jsx
@@ -4,7 +4,7 @@ import { Menu, Segment, Icon } from "semantic-ui-react";
 const DisplayHeader = () => {
   return (
     <Segment inverted>
-      <Menu inverted pointing secondary>
+      <Menu inverted pointing secondary >
         <Menu.Item
           name="main-header"
           id="main-header"
